@@ -9,18 +9,23 @@
     "@fortawesome/free-regular-svg-icons": "^5.10.1",
     "@fortawesome/free-solid-svg-icons": "^5.10.1",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.6.0",
     "font-awesome": "^4.7.0",
     "leaflet": "^1.5.1",
-    "react": "^16.8.6",
-    "react-bootstrap": "^1.0.0-beta.10",
-    "react-dom": "^16.8.6",
-    "react-ga": "^2.6.0",
+    "react": "^17.0.0",
+    "react-bootstrap": "^1.6.4",
+    "react-dom": "^17.0.0",
+    "react-ga": "^3.3.0",
     "react-leaflet": "^2.4.0",
     "react-router": "^5.0.1",
     "react-router-dom": "^5.0.1",
-    "react-scripts": "3.0.1",
-    "react-wordcloud": "^1.0.8"
+    "react-wordcloud": "^1.2.7"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/preset-react": "^7.16.7",
+    "gh-pages": "^3.2.3",
+    "react-scripts": "5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -45,8 +50,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "gh-pages": "^2.1.1"
   }
 }
